+++ tests/test_txmonitor.py
@@ -1,4 +1,4 @@
-"""Gas helpers."""
+"""Transaction monitoring tests."""
 import secrets
 
 import pytest
@@ -12,6 +12,7 @@ from web3._utils.transactions import fill_nonce
 
 from smart_contracts_for_testing.gas import estimate_gas_fees, GasPriceMethod, apply_gas
 from smart_contracts_for_testing.token import create_token
+from smart_contracts_for_testing.txmonitor import wait_transactions_to_complete
 
 
 @pytest.fixture
@@ -57,22 +58,8 @@ def hot_wallet(web3, hot_wallet_private_key) -> LocalAccount:
     return Account.from_key(hot_wallet_private_key)
 
 
-def test_gas_fees_london(web3: Web3, deployer: str):
-    """Estimate gas fees on London hard-fork compatible blockchain.
-
-    Note: We cannot test for non-London EVMs, as EthereumTester does not support them.
-
-    https://github.com/ethereum/eth-tester/issues/233
-    """
-    fees = estimate_gas_fees(web3)
-    assert fees.method == GasPriceMethod.london
-    assert fees.base_fee == 1000000000
-    assert fees.max_fee_per_gas == 3000000000
-    assert fees.max_priority_fee_per_gas == 1000000000
-
-
-def test_raw_transaction_with_gas(web3: Web3, eth_tester, deployer: HexAddress, hot_wallet: LocalAccount):
-    """Create a raw transaction with gas information."""
+def test_wait_txs_parallel(web3: Web3, eth_tester, deployer: HexAddress, hot_wallet: LocalAccount):
+    """Wait multiple transactions to complete in parallel."""
 
     gas_fees = estimate_gas_fees(web3)
 
@@ -83,20 +70,39 @@ def test_raw_transaction_with_gas(web3: Web3, eth_tester, deployer: HexAddress, 
     web3.eth.send_transaction({"from": deployer, "to": hot_wallet.address, "value": 1 * 10**18})
     token.functions.transfer(hot_wallet.address, 50_000 * 10**18).transact({"from": deployer})
 
+    nonce = web3.eth.get_transaction_count(hot_wallet.address)
+
     # Create a raw transaction
     # Move 10 tokens from deployer to user1
     # https://web3py.readthedocs.io/en/stable/contracts.html?highlight=buildTransaction#web3.contract.ContractFunction.buildTransaction
-    tx = token.functions.transfer(hot_wallet.address, 10 * 10**18).buildTransaction({
+    tx1 = token.functions.transfer(hot_wallet.address, 10 * 10**18).buildTransaction({
+        "from": hot_wallet.address,
+        'chainId': web3.eth.chain_id,
+        "gas": 150_000,  # 150k gas should be more than enough for ERC20.transfer(),
+        "nonce": nonce,
+    })
+
+    tx2 = token.functions.transfer(hot_wallet.address, 10 * 10**18).buildTransaction({
         "from": hot_wallet.address,
         'chainId': web3.eth.chain_id,
         "gas": 150_000,  # 150k gas should be more than enough for ERC20.transfer()
+        "nonce": nonce + 1,
     })
 
-    tx = fill_nonce(web3, tx)
-    apply_gas(tx, gas_fees)
+    apply_gas(tx1, gas_fees)
+    apply_gas(tx2, gas_fees)
 
-    signed = hot_wallet.sign_transaction(tx)
-    tx_hash = web3.eth.send_raw_transaction(signed.rawTransaction)
-    receipt = web3.eth.get_transaction_receipt(tx_hash)
-    assert receipt.status == 1   # 1=success and mined
+    signed1 = hot_wallet.sign_transaction(tx1)
+    signed2 = hot_wallet.sign_transaction(tx2)
+
+    tx_hash1 = web3.eth.send_raw_transaction(signed1.rawTransaction)
+    tx_hash2 = web3.eth.send_raw_transaction(signed2.rawTransaction)
+
+    complete = wait_transactions_to_complete(web3, [tx_hash1, tx_hash2])
+
+    # Check both transaction succeeded
+    for receipt in complete.values():
+        assert receipt.status == 1  # tx success
+
+